requirements.txt
pyyaml
click
glob